Fixing View and ViewModel
