--- dailyCollectionPlan/src/main/java/org/rmj/guanzongroup/ghostrider/dailycollectionplan/ViewModel/VMCNA.java
@@ -1,48 +1,0 @@
-package org.rmj.guanzongroup.ghostrider.dailycollectionplan.ViewModel;
-
-import android.app.Application;
-
-import androidx.annotation.NonNull;
-import androidx.lifecycle.AndroidViewModel;
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-
-import org.rmj.g3appdriver.GRider.Database.Entities.EBranchInfo;
-import org.rmj.g3appdriver.GRider.Database.Entities.EDCPCollectionDetail;
-import org.rmj.g3appdriver.GRider.Database.Repositories.RBranch;
-import org.rmj.g3appdriver.GRider.Database.Repositories.RDailyCollectionPlan;
-
-public class VMCNA extends AndroidViewModel {
-    public static final String TAG = VMCNA.class.getSimpleName();
-    private final RBranch poBranch;
-    private final RDailyCollectionPlan poDcp;
-
-    private final MutableLiveData<EDCPCollectionDetail> poDcpDetail = new MutableLiveData<>();
-    private final MutableLiveData<String> psTransNox = new MutableLiveData<>();
-    private final MutableLiveData<String> psEntryNox = new MutableLiveData<>();
-
-    public VMCNA(@NonNull Application application) {
-        super(application);
-        this.poBranch = new RBranch(application);
-        this.poDcp = new RDailyCollectionPlan(application);
-    }
-
-    public void setParameter(String TransNox, String EntryNox) {
-        this.psTransNox.setValue(TransNox);
-        this.psEntryNox.setValue(EntryNox);
-    }
-
-    public LiveData<EDCPCollectionDetail> getCollectionDetail(){
-        return poDcp.getCollectionDetail(psTransNox.getValue(), psEntryNox.getValue());
-    }
-
-    public void setCurrentCollectionDetail(EDCPCollectionDetail detail){
-        this.poDcpDetail.setValue(detail);
-    }
-
-    public LiveData<EBranchInfo> getUserBranchEmployee(){
-        return poBranch.getUserBranchInfo();
-    }
-
-
-}
